requirements.txt
ml-collections==0.1.1
pandas==2.2.2
Pillow==10.4.0
torch==2.3.0
torchvision==0.18.0
numpy==2.1.1
scipy==1.14.1
tqdm==4.66.5
timm==1.0.15